% TODO File path/AT.A.from.particle.no.Rd
\name{AT.A.from.particle.no}
\alias{AT.A.from.particle.no}
\title{AT.A.from.particle.no}
\description{Returns mass number for given particle number}
\usage{AT.A.from.particle.no(particle.no)
}
\arguments{
  \item{particle.no}{ particle index number (array of size n) (see also \code{\link{particle.no}}).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{A}{ mass number (array of size n)}
  \item{return}{return}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.4/src/AT_DataParticle.c#L53}
}


