% TODO File path/AT.atomic.weight.from.Z.Rd
\name{AT.atomic.weight.from.Z}
\alias{AT.atomic.weight.from.Z}
\title{AT.atomic.weight.from.Z}
\description{Returns atomic weight for given Z}
\usage{AT.atomic.weight.from.Z(Z)
}
\arguments{
  \item{Z}{ atomic number (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{atomic.weight}{ atomic weight (array of size n)}
  \item{return}{return}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.6.0/src/AT_Dat
aParticle.c#L109}
}
