% TODO File path/AT.average.A.from.composition.Rd
\name{AT.average.A.from.composition}
\alias{AT.average.A.from.composition}
\title{AT.average.A.from.composition}
\description{Computes the average mass number for a given material composition
}
\usage{AT.average.A.from.composition(A, weight.fraction)
}
\arguments{
  \item{A}{ mass numbers of constituents (array of size n).}
  \item{weight.fraction}{ relative fractions of weight of constituents (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{average.A}{ average A}
}


