% TODO File path/AT.density.g.cm3.from.element.acronym.Rd
\name{AT.density.g.cm3.from.element.acronym}
\alias{AT.density.g.cm3.from.element.acronym}
\title{AT.density.g.cm3.from.element.acronym}
\description{Returns the elemental densities in g/cm3 for given elemental
 symbols}
\usage{AT.density.g.cm3.from.element.acronym(acronym)
}
\arguments{
  \item{acronym}{ elemental symbols (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{density}{ corresponding elemental densities in g/cm3 (array of size n)}
  \item{status}{status}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/trunk/src/AT_DataPart
icle.c#L419}
}
