% TODO File path/AT.electron.density.m3.from.material.no.Rd
\name{AT.electron.density.m3.from.material.no}
\alias{AT.electron.density.m3.from.material.no}
\title{AT.electron.density.m3.from.material.no}
\description{Get electron density [1/m3] for materials}
\usage{AT.electron.density.m3.from.material.no(material.no)
}
\arguments{
  \item{material.no}{ material indices (array of size n) (see also
 \code{\link{material.no}}).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{electron.density.m3}{ electron densities per m3 (array of size n)}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/trunk/src/AT_DataMate
rial.c#L327}
}
