% TODO File path/AT.element.acronym.from.Z.Rd
\name{AT.element.acronym.from.Z}
\alias{AT.element.acronym.from.Z}
\title{AT.element.acronym.from.Z}
\description{Returns elemental symbols for given Z}
\usage{AT.element.acronym.from.Z(Z)
}
\arguments{
  \item{Z}{ atomic numbers (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{acronym}{ corresponding elemental symbols (array of size n)}
  \item{status}{status}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.4/src/AT_DataParticle.c#L355}
}


