% TODO File path/AT.Gauss.IDF.Rd
\name{AT.Gauss.IDF}
\alias{AT.Gauss.IDF}
\title{AT.Gauss.IDF}
\description{Compute Gauss inverse distribution function (for compatibility)
}
\usage{AT.Gauss.IDF(rnd)
}
\arguments{
  \item{rnd}{ random number from uniform distribution between 0 and 1 (array
 of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{lambda.gauss}{ resulting Gauss lambda (array of size n)}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.6.2/src/AT_Ene
rgyLoss.c#L584}
}
