% TODO File path/AT.Gauss.PDF.Rd
\name{AT.Gauss.PDF}
\alias{AT.Gauss.PDF}
\title{AT.Gauss.PDF}
\description{Computes Gauss probability density function (for compatibility)
}
\usage{AT.Gauss.PDF(lambda.gauss)
}
\arguments{
  \item{lambda.gauss}{ Gauss lambda (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{density}{ resulting density (array of size n)}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.6.3/src/AT_Ene
rgyLoss.c#L578}
}
