% TODO File path/AT.I.eV.from.element.acronym.Rd
\name{AT.I.eV.from.element.acronym}
\alias{AT.I.eV.from.element.acronym}
\title{AT.I.eV.from.element.acronym}
\description{Returns the I values in eV for given elemental symbol}
\usage{AT.I.eV.from.element.acronym(acronym)
}
\arguments{
  \item{acronym}{ elemental symbols (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{I}{ corresponding I values in eV (array of size n)}
  \item{status}{status}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.3/src/AT_DataParticle.c#L448}
}


