% TODO File path/AT.lambda.Landau.from.lambda.Vavilov.Rd
\name{AT.lambda.Landau.from.lambda.Vavilov}
\alias{AT.lambda.Landau.from.lambda.Vavilov}
\title{AT.lambda.Landau.from.lambda.Vavilov}
\description{Transforms lambda parameter according to Vavilov definition
 into Landau definition
}
\usage{AT.lambda.Landau.from.lambda.Vavilov(lambda.vavilov, kappa)
}
\arguments{
  \item{lambda.vavilov}{ lambda parameter acc. to Vavilov (array of size n).}
  \item{kappa}{ kappa parameter (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{lambda.landau}{ (array of size n)}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/trunk/src/AT_EnergyLo
ss.c#L311}
}
