% TODO File path/AT.lambda.Vavilov.from.lambda.Landau.Rd
\name{AT.lambda.Vavilov.from.lambda.Landau}
\alias{AT.lambda.Vavilov.from.lambda.Landau}
\title{AT.lambda.Vavilov.from.lambda.Landau}
\description{Transforms lambda parameter according to Landau definition
 into Vavilov definition
}
\usage{AT.lambda.Vavilov.from.lambda.Landau(lambda.landau, kappa)
}
\arguments{
  \item{lambda.landau}{ lambda parameter acc. to Landau (array of size n).}
  \item{kappa}{ kappa parameter (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{lambda.vavilov}{ (array of size n)}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/trunk/src/AT_EnergyLo
ss.c#L328}
}
