% TODO File path/AT.Moliere.function.f0.Rd
\name{AT.Moliere.function.f0}
\alias{AT.Moliere.function.f0}
\title{AT.Moliere.function.f0}
\description{Returns values f0(red_Theta) of the first Moliere function
}
\usage{AT.Moliere.function.f0(red.Theta)
}
\arguments{
  \item{red.Theta}{ reduced angle variable.}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{first}{first}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.3/src/AT_MultipleCoulombScattering.c#L217}
}


