% TODO File path/AT.Moliere.function.f1.Rd
\name{AT.Moliere.function.f1}
\alias{AT.Moliere.function.f1}
\title{AT.Moliere.function.f1}
\description{Returns values f1(red_Theta) of the second Moliere function
}
\usage{AT.Moliere.function.f1(red.Theta)
}
\arguments{
  \item{red.Theta}{ reduced angle variable.}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{second}{second}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.4/src/AT_MultipleCoulombScattering.c#L231}
}


