% TODO File path/AT.Moliere.function.f2.Rd
\name{AT.Moliere.function.f2}
\alias{AT.Moliere.function.f2}
\title{AT.Moliere.function.f2}
\description{Returns values f2(red_Theta) of the third Moliere function
}
\usage{AT.Moliere.function.f2(red.Theta)
}
\arguments{
  \item{red.Theta}{ reduced angle variable.}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{third}{third}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.4/src/AT_MultipleCoulombScattering.c#L255}
}


