% TODO File path/AT.nuclear.spin.from.particle.no.Rd
\name{AT.nuclear.spin.from.particle.no}
\alias{AT.nuclear.spin.from.particle.no}
\title{AT.nuclear.spin.from.particle.no}
\description{Returns nuclear spin from particle no}
\usage{AT.nuclear.spin.from.particle.no(particle.no)
}
\arguments{
  \item{particle.no}{ particle index number (array of size n) (see also \code{\link{particle.no}}).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{I}{ nuclear spin (array of size n)}
  \item{status}{status}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.2/src/AT_DataParticle.c#L184}
}


