% TODO File path/AT.Z.from.element.acronym.Rd
\name{AT.Z.from.element.acronym}
\alias{AT.Z.from.element.acronym}
\title{AT.Z.from.element.acronym}
\description{Returns Z for given elemental symbols}
\usage{AT.Z.from.element.acronym(acronym)
}
\arguments{
  \item{acronym}{ elemental symbols (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{Z}{ corresponding Z (array of size n)}
  \item{status}{status}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.4/src/AT_DataParticle.c#L326}
}


