% TODO File path/AT.Z.from.particle.no.Rd
\name{AT.Z.from.particle.no}
\alias{AT.Z.from.particle.no}
\title{AT.Z.from.particle.no}
\description{Returns atomic number for given particle number}
\usage{AT.Z.from.particle.no(particle.no)
}
\arguments{
  \item{particle.no}{ particle index number (array of size n) (see also
 \code{\link{particle.no}}).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{Z}{ atomic number (array of size n)}
  \item{return}{return}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.6.3/src/AT_Dat
aParticle.c#L81}
}
