% TODO File path/AT.atomic.weight.from.element.acronym.Rd
\name{AT.atomic.weight.from.element.acronym}
\alias{AT.atomic.weight.from.element.acronym}
\title{AT.atomic.weight.from.element.acronym}
\description{Returns atomic weight for given elemental symbols}
\usage{AT.atomic.weight.from.element.acronym(acronym)
}
\arguments{
  \item{acronym}{ elemental symbols (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{atomic.weight}{ corresponding A (array of size n)}
  \item{status}{status}
}
\seealso{
View the C source code here:

\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.5.3/src/AT_DataParticle.c#L391}
}


