% TODO File path/AT.Landau.PDF.Rd
\name{AT.Landau.PDF}
\alias{AT.Landau.PDF}
\title{AT.Landau.PDF}
\description{Computes the Landau probability density function using CERNLIB
 (G115)
}
\usage{AT.Landau.PDF(lambda.landau)
}
\arguments{
  \item{lambda.landau}{ Landau lambda (array of size n).}
}
\value{
% TODO proper return definition of lists!!! ADD
% NUMBER_OF_FIELD_COMPONENT_DESCRIBTION AGAIN!!!)
  \item{density}{ resulting density (array of size n)}
}
\seealso{
View the C source code here:
\url{http://sourceforge.net/apps/trac/libamtrack/browser/tags/0.6.2/src/AT_Ene
rgyLoss.c#L77}
}
